chore: remove dropped legacy tables from schema and code
- Remove registered_groups, router_state, sessions_legacy from
  _create_schema (tables already dropped in production DB)
- Remove get_router_state/set_router_state functions (no longer needed)
- Remove legacy table tests that reference dropped tables
- Keep migration helper functions (get_all_registered_groups, etc.)
  for migration script compatibility
- Migration test now creates legacy tables inline before testing
