+++ py/src/nanoclaw/db/__init__.py
@@ -34,9 +34,7 @@ from nanoclaw.db.pg import (
     get_last_group_sync,
     get_messages_since,
     get_new_messages_for_conversations,
-    get_registered_group,
     get_roles_for_tenant,
-    get_router_state,
     get_session,
     get_session_legacy,
     get_task_by_id,
@@ -47,7 +45,6 @@ from nanoclaw.db.pg import (
     log_task_run,
     set_last_group_sync,
     set_registered_group,
-    set_router_state,
     set_session,
     set_session_legacy,
     store_chat_metadata,
@@ -98,9 +95,7 @@ __all__ = [
     "get_last_group_sync",
     "get_messages_since",
     "get_new_messages_for_conversations",
-    "get_registered_group",
     "get_roles_for_tenant",
-    "get_router_state",
     "get_session",
     "get_session_legacy",
     "get_task_by_id",
@@ -111,7 +106,6 @@ __all__ = [
     "log_task_run",
     "set_last_group_sync",
     "set_registered_group",
-    "set_router_state",
     "set_session",
     "set_session_legacy",
     "store_chat_metadata",
